Updated the Parser
The parser was working for only one specific input file and I updated it to take in an input text file and updated the output file based on those. Also added documentation to the parser to explain why the output file is arranged the way it is and what the 10K xbrl extraction file does with it.