Finalized 10K extraction. Bugfix
There is was an issue with creating the subfolders if there was a / in the name of the company. In which case the subfolder would not correctly be created and it would break the code. This fixes this issues and updates the output. Have added the 2 files to reflect the fix.