Request Issue
There is a issue with the requests as there are over 900000 cik numbers and it is overloading the website even early one. Possible fixes are in this document. One way is to set multiple agents and randomize the agents and add a small timer in between each search. Added example with accessing the table inside with these above fixes. 10K xbrl would function in the same way.